Moved chunkArray function to common source file
